WIP: Attempt to move scripts
This commit will attempt to move scripts for the new button dialog into the correct files. This needs help from Roni to get it working properly.
